backend/pyproject.toml
[tool.poetry]
name = "sigdig-backend"
version = "0.1.0"
description = "Signal Digest backend API"
authors = ["Signal Digest Team"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.11,<3.14"
fastapi = "^0.115.5"
uvicorn = {extras = ["standard"], version = "^0.32.1"}
python-multipart = "^0.0.17"
sqlalchemy = "^2.0.36"
alembic = "^1.14.0"
psycopg2-binary = "^2.9.10"
python-dotenv = "^1.0.1"
pydantic = "^2.10.3"
pydantic-settings = "^2.6.1"
langchain = "^1.2.3"
langchain-google-genai = "^2.0.5"
langchain-core = "^1.2.6"
feedparser = "^6.0.11"
requests = "^2.32.3"
python-dateutil = "^2.9.0"

[tool.poetry.group.dev.dependencies]
black = "^24.0.0"
ruff = "^0.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
target-version = "py311"
